change: separate FrameGraph into its own crate
