+++ crates/renderer/src/passes/atmosphere.rs
@@ -1,16 +1,15 @@
 use anyhow::Result;
+use gfx::state::RenderState;
 use glam::{Mat4, Vec3Swizzles, Vec4};
 use hot_reload::IntoDynamic;
 use inject::DI;
+use pass::FrameGraph;
 use ph::vk;
 use phobos as ph;
 use phobos::{Allocator, GraphicsCmdBuffer};
 use scheduler::EventBus;
 use statistics::{RendererStatistics, TimedCommandBuffer};
 use world::World;
-
-use crate::util::graph::FrameGraph;
-use crate::world_renderer::RenderState;
 
 /// The atmosphere renderer is responsible for rendering the
 /// atmosphere into the frame graph.
