+++ crates/renderer/src/passes/terrain.rs
@@ -1,18 +1,17 @@
 use anyhow::Result;
 use assets::storage::AssetStorage;
+use gfx::state::RenderState;
 use gfx::{create_linear_sampler, create_raw_sampler};
 use glam::{Mat4, Vec3};
 use hot_reload::IntoDynamic;
 use inject::DI;
+use pass::FrameGraph;
 use ph::vk;
 use phobos::prelude as ph;
 use phobos::prelude::traits::*;
 use scheduler::EventBus;
 use statistics::{RendererStatistics, TimedCommandBuffer};
 use world::World;
-
-use crate::util::graph::FrameGraph;
-use crate::world_renderer::RenderState;
 
 /// The terrain renderer. Stores resources it needs for rendering.
 /// This struct renders the main terrain mesh.
