+++ crates/renderer/src/util/mod.rs
@@ -1,2 +1,1 @@
-pub mod graph;
 pub mod targets;
