+++ crates/renderer/src/world_renderer.rs
@@ -1,10 +1,12 @@
 use anyhow::Result;
 use camera::CameraState;
+use gfx::state::RenderState;
 use gfx::SharedContext;
 use glam::{Mat3, Mat4, Vec3};
 use gui::util::image_provider::ImageProvider;
 use hot_reload::IntoDynamic;
 use inject::DI;
+use pass::FrameGraph;
 use phobos::{vk, PassBuilder, PhysicalResourceBindings, PipelineBuilder, VirtualResource};
 use scheduler::EventBus;
 use world::World;
@@ -13,31 +15,7 @@ use crate::passes::atmosphere::AtmosphereRenderer;
 use crate::passes::terrain::TerrainRenderer;
 use crate::postprocess::tonemap::Tonemap;
 use crate::ui_integration::UIIntegration;
-use crate::util::graph::FrameGraph;
 use crate::util::targets::{RenderTargets, SizeGroup};
-
-/// Stores world state in a format that the renderer needs, such as
-/// normalized direction vectors instead of rotations,
-/// camera view and projection matrices, etc.
-#[derive(Debug, Default)]
-pub struct RenderState {
-    /// Camera view matrix
-    pub view: Mat4,
-    /// Camera projection matrix
-    pub projection: Mat4,
-    /// Premultiplied `projection * view` matrix
-    pub projection_view: Mat4,
-    /// Inverse of the projection matrix
-    pub inverse_projection: Mat4,
-    /// Inverse of `projection * view`
-    pub inverse_projection_view: Mat4,
-    /// Inverse of the camera's view matrix with the translation component removed
-    pub inverse_view_rotation: Mat4,
-    /// Direction vector pointing away from the sun
-    pub sun_direction: Vec3,
-    /// Camera position in world space
-    pub cam_position: Vec3,
-}
 
 /// The world renderer is responsible for all the rendering logic
 /// of the scene.
