+++ crates/world/Cargo.toml
@@ -3,8 +3,6 @@ name = "world"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 anyhow = "1.0.70"
 glam = "0.24.0"
